requirements.txt
streamlit
flask
pandas
snowflake-connector-python
